apply context_api react
