+++ package.json
@@ -13,6 +13,7 @@
     "bootstrap": "^5.3.1",
     "eslint-config-react-app": "^7.0.1",
     "firebase": "^10.4.0",
+    "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-bootstrap": "^2.8.0",
     "react-dom": "^18.2.0",
