+++ src/App.jsx
@@ -1,11 +1,12 @@
 import {createBrowserRouter, RouterProvider} from 'react-router-dom';
 import Pedidos from './pages/pedidos';
 import Estoque from "./pages/estoque"
+import PedidoContextProvider from './contexts/pagePedidos';
 
 const browser = createBrowserRouter([ //Criação da estrutura de navegação.
   {
     path:'/',
-    element:<Pedidos/>
+    element:<PedidoContextProvider><Pedidos/></PedidoContextProvider>
   },
   {
     path:'/estoque',
